refactor(map-next): unify farm and initiative profiles
FarmProfile and InitiativeProfile were ~85% the same file: identical
superform setup, drawer header, read/edit switch, save-bar error mapping,
sticky cancel/save, share action and contact CTA, differing only in which
domain sections they render.

Extract that scaffolding into EntryProfile, which owns the form and passes
it to `editSections`/`readSections` snippets. FarmProfile and
InitiativeProfile keep their public props and become thin wrappers that
supply their own sections, so MapSidebar is unchanged.

Layout drift between the two is resolved: initiatives adopt the farm's p-2
drawer header and footer padding, and the `divide-y` that farms drew
between edit-mode sections is dropped, so both types now match read mode —
which has never drawn dividers on either. That restores the read/edit
"divider rhythm" parity the section comment claims (F4.2). The only
remaining per-type difference besides the sections themselves is the
initiative intro copy, which is real content with no farm equivalent.

svelte-check, eslint, 130 unit tests and the full 87-test Playwright suite
(including both edit-parity tests) pass.
